+++ run/userland/01_second.c
@@ -3,5 +3,5 @@
 
 void _start() {
     console_puts("[user] Second userland app\n");
-    for (;;) __asm__("hlt");
+    return;
 }